requirements.txt
click
requests
ruamel.yaml
appdirs

bs4
html2text